2018-08-06  Richard Larsson  <larsson@mps.mpg.de>
	* arts-2-3-1089
	
        Updates of fast-wigner calculations by H. Johansson,
	should now support multithreading, though this has not been
	adapted in ARTS-code yet.
	
	The rest are just small code dumps for some ongoing work on
	line mixing calculations.
	
	*  3rdparty/relmat/module_linemixing.F90:  Fix to hard-coded
	temperature dependency
	
	* src/CMakeLists.txt:  Added a linemixing.cc file
	
	* src/abs_species_tags.cc,h:  Added SpeciesMass,
	IsSpecies, and IsIsotopologue to SpeciesTag-class for
	easier use without having to access the species-mapping
	itself (use these on own risk, because the ARTS species-
	mapping is riskful to use for these purposes unless you
	are absolutely certain what you can have available)
	
	* src/complex.h: comments update from the copy-paste.
	
	* src/linefunctions.cc,h:  Update so function takes reference
	instead of copies
	
	* src/linemixing.cc,h:  Added.  Contains an experimental 
	re-implementation of the simplest case of relmat.  Not
	for real use yet...
	
	* src/m_linemixing.cc: TestLineMixing to test the above
	
	* src/methods.cc: as above
	
	* src/rational.cc,h:  Added toInt() because Wigner-code
	needs integers
	
	* src/wigner_functions.cc,h: made co2_ecs_wigner_symbol public



git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@11043 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ 3rdparty/wigner/fastwigxj/src/gen_3j.cc
@@ -66,7 +66,7 @@ void loop_four(stats &st,
       uint64_t key =
 	wigner3j_regge_canonicalise(jv);
       
-      key &= ~1; // ignore the sign bit
+      key &= ~((uint64_t) 1); // ignore the sign bit
       
       insert_buf(key);
     }
